% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_document_terms.R
\name{clean_abstracts}
\alias{clean_abstracts}
\title{Clean abstract text}
\usage{
clean_abstracts(abstracts)
}
\arguments{
\item{abstracts}{A data.frame containing a column with unique identifiers in
a column named "ut", and another
column with abstract text (as a character vector).}
}
\value{
A data.frame.
}
\description{
This function takes a tibble as input with a document identifier column, and
another column containing document text, and outputs a tokenized clean
data.frame with word stems.
}
