% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstracts.R
\name{get_abstracts}
\alias{get_abstracts}
\title{Get abstracts for papers matching search terms}
\usage{
get_abstracts(terms)
}
\arguments{
\item{terms}{A character vector of search terms to use on Web of Science.}
}
\value{
A tibble with one row per publication matching search terms, with
abstracts in one column.
}
\description{
Get abstracts for papers matching search terms
}
