% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_document_terms.R
\name{get_document_terms}
\alias{get_document_terms}
\title{Get a document term matrix from a set of abstracts (or other text)}
\usage{
get_document_terms(abstracts)
}
\arguments{
\item{abstracts}{A data.frame containing a column with DOIs, and another
column with abstract text (as a character vector).}
}
\value{
An object of class \code{tm::DocumentTermMatrix}.
}
\description{
This function takes a tibble as input with a document identifier column, and
another column containing document text, and outputs a document-term
matrix that can be used for downstream analysis.
}
