% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstracts.R
\name{get_abstracts}
\alias{get_abstracts}
\title{Get abstracts for papers in a .bib file}
\usage{
get_abstracts(bibfile)
}
\arguments{
\item{bibfile}{Path to a .bib file (character vector)}
}
\value{
A data frame with one row per DOI, and a column containing abstract
text (if available).
}
\description{
By default abstracts are acquired via crossref and the fulltext package
under the hood (it may make sense to change this in the future...)
}
