% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_document_terms.R
\name{get_document_terms}
\alias{get_document_terms}
\title{Get a document term matrix from a set of abstracts (or other text)}
\usage{
get_document_terms(abstracts)
}
\arguments{
\item{abstracts}{A data.frame containing a column with unique identifiers in
a column named "ut", and another called "word_stem".}
}
\value{
An object of class \code{DocumentTermMatrix}.
}
\description{
This function takes a tibble as input with tokenized text, e.g., as
created by the \code{clean_abstracts} function, and generates a document
term matrix using word stems.
}
